requirements.txt
Appium-Python-Client==2.11.1
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
exceptiongroup==1.1.1
func-timeout==4.3.5
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
loguru==0.7.0
outcome==1.2.0
packaging==23.1
playwright==1.35.0
pluggy==1.2.0
poium==1.3.2
pyee==9.0.4
PySocks==1.7.1
pytest==7.4.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.7.0
urllib3==2.0.3
wsproto==1.2.0
